feat: extract url builder
