+++ src/models/pages/DetailsRequest.tsx
@@ -1,0 +1,4 @@
+export type DetailsRequest = {
+  username: string;
+  repo: string;
+};
