+++ src/models/pages/DetailsResponse.tsx
@@ -1,0 +1,7 @@
+import { GithubRepository } from '../GithubRepository';
+import { GithubErrorResponse } from '../GithubErrorResponse';
+
+export type DetailsResponse = {
+  details: GithubRepository | null;
+  error: GithubErrorResponse | null;
+};
