+++ src/models/pages/PageRequest.tsx
@@ -1,0 +1,7 @@
+import { ParsedUrlQuery } from 'querystring';
+
+export interface PageRequest extends ParsedUrlQuery {
+  page?: string;
+  count?: string;
+  query?: string;
+}
